# api.yaml
openapi: 3.1.1
info:
  title: Doctree API
  description: API для иерархической системы хранения документов
  version: "2.0.0"
servers:
  - url: /api/v2
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
      required: [username, password, email]
    SchedulerRegisterRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Авторизация по логину и паролю
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: Успешный вход
        "400":
          description: Неверный логин или пароль
  /auth/logout:
    post:
      tags: [Auth]
      summary: Авторизация по логину и паролю
      responses:
        "200":
          description: Успешный выход
        "401":
          description: Пользователь не аутентифицирован
  /auth/register:
    post:
      tags: [Auth]
      summary: Регистрация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Пользователь создан
        "403":
          description: У пользователя нет прав администратора
        "400":
          description: Ошибка запроса
  /auth/scheduler:
    post:
      security: []   # создание планировщика без авторизации
      tags: [Auth]
      summary: Создать планировщик парсера
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchedulerRegisterRequest"
      responses:
        "201":
          description: Планировщик создан
        "403":
          description: У пользователя нет прав администратора
        "400":
          description: Ошибка запроса
...